refactor(test): move shared test helper into tests/common
Extract TestLog, setup_logging, create_context, testing_setup and
read_block_data from tests/test.rs into tests/common so upcoming
integration tests can reuse them. No behaviour change.
